+++ tflite2onnx/op/concat.py
@@ -2,7 +2,7 @@ import logging
 import tflite
 
 from tflite2onnx.op.activation import handleFusedActivation
-from tflite2onnx.op.operator import Operator
+from tflite2onnx.op.common import Operator
 
 logger = logging.getLogger('tflite2onnx')
