+++ tflite2onnx/op/padding.py
@@ -2,7 +2,7 @@ import logging
 import numpy as np
 import tflite
 
-from tflite2onnx.op.operator import Operator
+from tflite2onnx.op.common import Operator
 
 logger = logging.getLogger('tflite2onnx')
