+++ tflite2onnx/op/softmax.py
@@ -1,7 +1,7 @@
 import logging
 import tflite
 
-from tflite2onnx.op.operator import Operator
+from tflite2onnx.op.common import Operator
 
 logger = logging.getLogger('tflite2onnx')
